Hand-roll parallel optimization
A lot of work but I think it's looking quite nice. Still needs to be integrated more fully into the incremental compilation. Roughly the same speed as serial whole-module optimization, but will be faster when made incremental, and especially when there's actually a lot of parallelism in the dependency graph to exploit.