Add buffer indexation with tuples
my_values = my_buf.[my_indices];

Yet to add the corresponding store syntax, e.g. my_buf.[my_tuple] = values_tuple.

When the indexing argument is a ramp literal, it compiles directly to a vector load.

When the resultant tuple is vector-like but the argument is not a ramp, it compiles to a gather.

Can be improved still further to cases where the argument is a ramp but is not a ramp literal.